+++ package.json
@@ -165,13 +165,14 @@
   },
   "license": "GPL-3.0-or-later",
   "dependencies": {
-    "@gnosis.pm/dex-js": "^0.11.0",
     "@gnosis.pm/cow-runner-game": "^0.2.9",
+    "@gnosis.pm/dex-js": "^0.11.0",
     "@gnosis.pm/gp-v2-contracts": "^0.0.1-alpha.20",
     "@uniswap/default-token-list": "^2.0.0",
     "@web3-react/walletconnect-connector": "^6.2.0",
     "paraswap": "^4.18.0",
     "react-appzi": "^1.0.4",
+    "react-inlinesvg": "^2.3.0",
     "react-router-hash-link": "^2.4.0"
   }
 }
